refactor: move projects query into a composable
