+++ subtasks/tsconfig.app.json
@@ -8,7 +8,7 @@
     "paths": {
       "@annotations/*": ["./src/annotations/*"],
       "@components/*": ["./src/components/*"],
-      "@actions/*": ["./src/actions/*"],
+      "@actions/*": ["src/composables/*"],
     }
   }
 }
